使用web下载依赖测试

+++ setup.sh
@@ -66,22 +66,51 @@ if [ "$ARCH" = "x86_64" ]; then
     install_docker
     run_docker_compose
 elif [ "$ARCH" = "loongarch64" ]; then
-    yum install -y wget git
+    yum install -y wget
+    # # 安装 docker ce
+    # tar -xf ./installer/loongarch64/docker-27.0.3.tgz
+    # mv docker/* /usr/local/bin/
+    # rm -rf docker
+    # cp ./installer/loongarch64/docker.service /etc/systemd/system/docker.service
+    # systemctl daemon-reload
+    # systemctl start docker
+    # systemctl enable docker
+    # # 安装 docker compose
+    # cp ./installer/loongarch64/docker-compose-linux-loongarch64 /usr/local/bin/docker-compose
+    # chmod +x /usr/local/bin/docker-compose
+    if [ ! -f '/etc/systemd/system/docker.service' ]; then
     # 安装 docker ce
-    tar -xf ./installer/loongarch64/docker-27.0.3.tgz
-    mv docker/* /usr/local/bin/
-    rm -rf docker
-    cp ./installer/loongarch64/docker.service /etc/systemd/system/docker.service
-    systemctl daemon-reload
-    systemctl start docker
-    systemctl enable docker
-    # 安装 docker compose
-    cp ./installer/loongarch64/docker-compose-linux-loongarch64 /usr/local/bin/docker-compose
-    chmod +x /usr/local/bin/docker-compose
+      wget https://lib.storage.pigeon-server.cn/docker-27.0.3.tgz
+      # tar -xf docker-27.0.3.tgz
+      # cp docker/* /usr/local/bin/
+      # rm -rf docker
+      # 创建临时文件并获取文件列表
+      tempfile=$(mktemp)
+      tar -xvf docker-27.0.3.tgz | tee "$tempfile"
+      # 将文件复制到 /usr/local/bin/
+      sudo cp docker/* /usr/local/bin/
+      # 从文件列表中创建软链接
+      while read -r file; do
+          filename=$(basename "$file")
+          sudo ln -s /usr/local/bin/"$filename" /usr/bin/"$filename"
+      done < "$tempfile"
+      # 删除临时文件
+      rm "$tempfile"
+      rm docker-27.0.3.tgz
+      # 删除解压的文件夹
+      rm -rf docker
+      wget https://lib.storage.pigeon-server.cn/docker.service -O /etc/systemd/system/docker.service
+      systemctl start docker
+      systemctl enable docker
+    fi
+    if [ ! -f '/usr/local/bin/docker-compose' ]; then
+      # 安装 docker compose
+      wget -O /usr/local/bin/docker-compose https://lib.storage.pigeon-server.cn/docker-compose-linux-loongarch64
+      chmod +x /usr/local/bin/docker-compose
+      ln -s /usr/local/bin/docker-compose /usr/bin/docker-compose
+    fi
     # 运行docker
     docker-compose -f docker-compose.loongarch.yml up -d
-    # 清理
-    rm -rf docker
 else
     echo "不支持的系统架构: ${ARCH}" && exit 1;
 fi
